수정 삭제 버튼 출력 조건문 추가

+++ admin/teachers/teachers_board/t_read.php
@@ -101,7 +101,7 @@ switch ($category) {
     <a href="<?=$redirect_url?>" class="btn btn-secondary">목록</a>
     <a href="t_like_up.php?pid=<?=$pid?>&category=<?=$category?>" class="btn btn-info">추천</a>
     <?php if (isset($_SESSION['TUID']) && $_SESSION['TUID'] == $data->user_id) : ?>
-    <!-- 수정/삭제 버튼 (본인이 작성한 글일 때만 표시) -->
+    <!-- 수정 삭제 버튼 본인이 작성한 글일 때만 표시 -->
     <a href="t_board_modify.php?pid=<?=$pid?>&category=<?=$category?>" class="btn btn-primary">수정</a>
     <a href="t_delete.php?pid=<?=$pid?>&category=<?=$category?>" class="btn btn-danger">삭제</a>
     <?php endif; ?>
@@ -141,11 +141,15 @@ switch ($category) {
           <?=$data->content?>
           </div>
           <div class="controls d-flex justify-content-end gap-1">
+          <?php if (isset($_SESSION['TUID']) && $_SESSION['TUID'] == $data->user_id) : ?>
+          <!-- 수정 삭제 버튼 본인이 작성한 댓글일 때만 표시 -->
             <button class="btn btn-primary sm" data-bs-toggle="modal" data-bs-target="#reply_edit<?=$data->pid?>">수정</button>
             <a href="t_reply_delete.php?pid=<?=$data->pid?>&b_pid=<?=$data->b_pid?>&category=<?=$category?>" class="btn btn-danger sm">삭제</a>
+          <?php endif; ?>
           </div>
         </div>
         <!-- //댓글 출력 부분 -->
+         
         <!-- modal -->
         <div class="modal fade" id="reply_edit<?=$data->pid?>" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
           <div class="modal-dialog">
